model_to_server/requirements.txt
fastapi
uvicorn
pillow
torch
ultralytics>=8.0.0
python-multipart